Add TypeScript foundations
