+++ lib/javascript/mdc/src/index.ts
