+++ lib/javascript/utils/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./typed";
+export * from "./untyped";
