+++ lib/javascript/utils/src/typed.ts
@@ -1,5 +1,3 @@
-import React from "react";
-
 // used in orchest-webserver and mdc-components only
 export function uuidv4() {
   return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, function (c) {
@@ -106,90 +104,6 @@ export function absoluteToRelativePath(path, cwd) {
   return relativePath;
 }
 
-// used in orchest-webserver only
-export function makeCancelable(promise, promiseManager) {
-  let hasCanceled_ = false;
-
-  let cancelablePromise = {
-    cancel() {
-      hasCanceled_ = true;
-    },
-  };
-
-  const wrappedPromise = new Promise((resolve, reject) => {
-    promise.then(
-      (val) => {
-        hasCanceled_ ? reject({ isCanceled: true }) : resolve(val);
-
-        promiseManager.clearCancelablePromise(cancelablePromise);
-      },
-      (error) => {
-        hasCanceled_ ? reject({ isCanceled: true }) : reject(error);
-
-        promiseManager.clearCancelablePromise(cancelablePromise);
-      }
-    );
-  });
-
-  cancelablePromise.promise = wrappedPromise;
-
-  promiseManager.appendCancelablePromise(cancelablePromise);
-
-  return cancelablePromise;
-}
-
-// used in orchest-webserver and mdc-components only
-export class RefManager {
-  constructor() {
-    this._refs = {};
-
-    this.refs = new Proxy(this._refs, {
-      get: (target, name, receiver) => {
-        if (!target[name]) {
-          return;
-        }
-        return target[name].current;
-      },
-      set: (target, name, value, receiver) => {
-        target[name] = value;
-      },
-    });
-
-    this.nrefs = new Proxy(this._refs, {
-      get: (target, name, receiver) => {
-        if (!target[name]) {
-          target[name] = new React.createRef();
-        }
-
-        return target[name];
-      },
-    });
-  }
-}
-
-// used in orchest-webserver only
-export class PromiseManager {
-  constructor() {
-    this.cancelablePromises = [];
-  }
-
-  appendCancelablePromise(cancelablePromise) {
-    this.cancelablePromises.push(cancelablePromise);
-  }
-
-  cancelCancelablePromises() {
-    for (let cancelablePromise of this.cancelablePromises) {
-      cancelablePromise.cancel();
-    }
-  }
-
-  clearCancelablePromise(cancelablePromise) {
-    let index = this.cancelablePromises.indexOf(cancelablePromise);
-    this.cancelablePromises.splice(index, 1);
-  }
-}
-
-// used in mdc-components only
 export function someParentHasClass(element, classname) {
   if (element.classList && element.classList.contains(classname)) return true;
   return (
@@ -280,7 +194,7 @@ export function arraysEqual(a, b) {
 }
 
 // used in orchest-webserver only
-export function makeRequest(method, url, body, onprogressCallback, timeout) {
+export function makeRequest(method, url, body?, onprogressCallback?, timeout?) {
   return new Promise(function (resolve, reject) {
     var xhr = new XMLHttpRequest();
     xhr.open(method, url);
@@ -346,6 +260,8 @@ export function makeRequest(method, url, body, onprogressCallback, timeout) {
 
 // used in orchest-webserver only
 export class PersistentLocalConfig {
+  localStoragePrefix: string;
+
   constructor(localStoragePrefix) {
     this.localStoragePrefix = localStoragePrefix;
   }
