+++ lib/javascript/utils/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../../tsconfig.json",
+  "exclude": ["node_modules"],
+  "include": ["src/**/*.ts"],
+}
