+++ services/orchest-webserver/client/package.json
@@ -7,7 +7,8 @@
   "scripts": {
     "build": "vite build",
     "dev": "vite",
-    "serve": "vite preview"
+    "serve": "vite preview",
+    "lint": "tsc --noEmit"
   },
   "dependencies": {
     "@orchest/lib-mdc": "workspace:*",
@@ -29,10 +30,12 @@
     "xterm-for-react": "^1.0.4"
   },
   "devDependencies": {
+    "@types/react": "17.0.3",
     "@vitejs/plugin-react-refresh": "1.3.1",
     "esbuild": "0.8.44",
     "npm-watch": "0.7.0",
     "sass": "1.32.8",
-    "vite": "2.1.5"
+    "vite": "2.1.5",
+    "typescript": "4.2.4"
   }
 }
